# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VoidScan Security Operations Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /threat/url-reputation:
    post:
      operationId: threatUrlReputation
      tags: [threat]
      summary: URL Reputation Checker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: URL reputation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReputationResult"
  /threat/ip-reputation:
    post:
      operationId: threatIpReputation
      tags: [threat]
      summary: IP Reputation Checker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: IP reputation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReputationResult"
  /threat/domain-reputation:
    post:
      operationId: threatDomainReputation
      tags: [threat]
      summary: Domain Reputation Checker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Domain reputation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReputationResult"
  /threat/phishing:
    post:
      operationId: threatPhishing
      tags: [threat]
      summary: Phishing Detection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Phishing analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhishingResult"
  /threat/suspicious-url:
    post:
      operationId: threatSuspiciousUrl
      tags: [threat]
      summary: Suspicious URL Analyzer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Suspicious URL analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuspiciousUrlResult"
  /threat/virustotal:
    post:
      operationId: threatVirusTotal
      tags: [threat]
      summary: VirusTotal Integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: VirusTotal scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirusTotalResult"
  /threat/abuseipdb:
    post:
      operationId: threatAbuseIpdb
      tags: [threat]
      summary: AbuseIPDB Integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: AbuseIPDB result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbuseIpdbResult"
  /threat/openphish:
    post:
      operationId: threatOpenPhish
      tags: [threat]
      summary: OpenPhish Feed Integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: OpenPhish check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResult"
  /threat/breach:
    post:
      operationId: threatBreach
      tags: [threat]
      summary: Known Data Breach Checker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Breach check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreachResult"

  # ─── File Analysis ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TargetInput:
      type: object
      required: [target]
      properties:
        target:
          type: string
...
    GenericResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
...
    ReputationResult:
      type: object
      required: [success, target]
      properties:
        success:
          type: boolean
        target:
          type: string
        score:
          type: ["integer", "null"]
        verdict:
          type: ["string", "null"]
        categories:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            type: object
            additionalProperties: true
        error:
          type: ["string", "null"]
    PhishingResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        url:
          type: ["string", "null"]
        isPhishing:
          type: ["boolean", "null"]
        confidence:
          type: ["number", "null"]
        indicators:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    SuspiciousUrlResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        url:
          type: ["string", "null"]
        riskScore:
          type: ["integer", "null"]
        riskLevel:
          type: ["string", "null"]
        flags:
          type: array
          items:
            type: string
        details:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
    VirusTotalResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        target:
          type: ["string", "null"]
        positives:
          type: ["integer", "null"]
        total:
          type: ["integer", "null"]
        scanDate:
          type: ["string", "null"]
        permalink:
          type: ["string", "null"]
        scans:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
        requiresApiKey:
          type: ["boolean", "null"]
    AbuseIpdbResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        ip:
          type: ["string", "null"]
        abuseConfidenceScore:
          type: ["integer", "null"]
        totalReports:
          type: ["integer", "null"]
        lastReportedAt:
          type: ["string", "null"]
        countryCode:
          type: ["string", "null"]
        isp:
          type: ["string", "null"]
        categories:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
        requiresApiKey:
          type: ["boolean", "null"]
    BreachResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        email:
          type: ["string", "null"]
        breaches:
          type: array
          items:
            type: object
            additionalProperties: true
        breachCount:
          type: ["integer", "null"]
        found:
          type: ["boolean", "null"]
        error:
          type: ["string", "null"]
...